Intelligently detect the minimum input size needed
Previously there was a hard-coded check to ensure the input image
was at least 1024x1024. But there are often cases where that size
of image is not needed (e.g. when adding an alternate iOS app icon
which doesn't need a corresponding marketing version)

Now the check is done based on the output sizes actually requested.

In addition, there is now an "allow_upsampling" flag so that this
check can be bypassed entirely.